% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databrary_list_data_files.R
\name{databrary_list_data_files}
\alias{databrary_list_data_files}
\title{List contents of materials folder on data volume}
\usage{
databrary_list_data_files(session_id = 51539)
}
\arguments{
\item{session_id}{Session ID for materials folder; default is 51539.}

\item{vol_id}{Databrary volume ID; default is 1280.}
}
\description{
List contents of materials folder on data volume
}
