% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databrary_load_data_csv.R
\name{databrary_load_data_csv}
\alias{databrary_load_data_csv}
\title{Loads CSV data file from Databrary}
\usage{
databrary_load_data_csv(session_id = 51539, asset_id = 380333, vb = FALSE)
}
\arguments{
\item{session_id}{Session ID (folder) for data files; default is 51539.}

\item{asset_id}{Asset ID (file); default is 380333.}

\item{vb}{Show verbose feedback; default is FALSE.}
}
\description{
Loads CSV data file from Databrary
}
