% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databrary_test_login.R
\name{databrary_test_login}
\alias{databrary_test_login}
\title{Store Databrary credentials in secure system keychain}
\usage{
databrary_test_login(env_var = "DATABRARY_LOGIN")
}
\arguments{
\item{databrary_user_id}{A valid Databrary user ID (email).}
}
\description{
Store Databrary credentials in secure system keychain
}
