% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databrary_verify_volume_access.R
\name{databrary_verify_volume_access}
\alias{databrary_verify_volume_access}
\title{Verify access to PLAY data volume on Databrary.}
\usage{
databrary_verify_volume_access(vol_id = 1280)
}
\arguments{
\item{vol_id}{Databrary volume ID; defaults to 1280.}
}
\description{
Verify access to PLAY data volume on Databrary.
}
